fix(desktop): keep live-view queries fresh across focus returns
The managed-agent log and workflow-runs queries are detail-view surfaces whose only off-focus refresh is conditional polling, so the 5-minute stale gate hid remote run starts and live log updates on refocus. Gate them at one poll tick (30s/10s) instead — neither feeds the app-wide focus storm. Also extracts the agent snapshot hooks into snapshotHooks.ts to keep agents/hooks.ts under the file-size ceiling, and replaces constant-assertion focus tests with behavioral ones.
